\name{WFO.example}
\alias{WFO.example}
\docType{data}
\title{World Flora Online (WFO) taxonomic backbone example data set}
\description{
This data set is a subset of the World Flora Online taxonomic backbone that allows running the first set of examples. In practical applications, users should first download a static copy of the Taxonomic Backbone data from \url{http://www.worldfloraonline.org/downloadData} .
}
\usage{data(WFO.example)}

\source{
World Flora Online. An Online Flora of All Known Plants.
\url{http://www.worldfloraonline.org}
}

\examples{
data(WFO.example)
}
\keyword{datasets}
